(FACT-3006) refactor internal_fact_manager.rb
Removed fact_augmenter.rb and moved the logic to
internal_fact_manager.rb
